+++ view/frontend/web/js/view/turnstile.js
@@ -8,7 +8,7 @@
 /*global define*/
 define(
     [
-        'PixelOpen_CloudflareTurnstile/js/view/Component',
+        'PixelOpen_CloudflareTurnstile/js/view/component',
         'Magento_Customer/js/customer-data'
     ],
     function (
@@ -31,7 +31,7 @@ define(
                 }
 
                 return this._super();
-            },
+            }
         });
     }
 );